+++ src/components/LayerMarkers.js
@@ -27,9 +27,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-//----------------------------------------------------------------//
-// Top Level Modules
-//----------------------------------------------------------------//
 import React from 'react'
 import {
   Marker,
@@ -48,19 +45,20 @@ import Button from '@material-ui/core/Button'
 import { makeStyles } from '@material-ui/core/styles'
 
 //----------------------------------------------------------------//
-// Custom Components
+// Hawg View Functions
 //----------------------------------------------------------------//
 import { render9line, render15line } from '../functions/renderData'
+import { distanceAndHeading } from '../functions/mathFunctions'
 
 //----------------------------------------------------------------//
 // Layer Markers Component
 //----------------------------------------------------------------//
 const LayerMarkers = props => {
 
-  const handleClose = () => {
+  /*const handleClose = () => {
     props.setActiveDialog(null)
     props.setFocusedMarker(null)
-  }
+  }*/
 
   const computedSize = props.markerSize * props.mapZoom
 
@@ -106,17 +104,32 @@ const LayerMarkers = props => {
     } catch (e) {
       position = `${marker.latlng.lat.toFixed(4)}, ${marker.latlng.lng.toFixed(4)}`
     }
+
+    let fromBE = null
+    if (props.anchor.id !== null) {
+      fromBE = distanceAndHeading(props.anchor.latlng, marker.latlng, props.anchor.declination)
+    }
     return (
-      <React.Fragment>
-        {marker.title}
-        <br />
-        {position}
-        <br />
-        {(marker.elevation !== 'Pending' && marker.elevation !== 'Elevation not found') ?
-          `${marker.elevation} feet`
-          : 'No elevation'
-        }
-      </React.Fragment>
+      <table>
+        <tbody>
+          <tr>
+            <td>{marker.title}</td>
+          </tr>
+          <tr>
+            <td>{position}</td>
+          </tr>
+          {fromBE !== null ? (
+            <tr>
+              <td>{props.anchor.name} {Number.parseInt(fromBE.heading)}&deg; / {fromBE.nm.toFixed(2)} NM</td>
+            </tr>
+          )
+            :
+            null}
+          <tr>
+            <td>{marker.elevation !== 'Pending' && marker.elevation !== 'Elevation not found' ? `${marker.elevation} feet` : 'No elevation'} </td>
+          </tr>
+        </tbody>
+      </table>
     )
   }